Can now save files as well!
It asks whether the user wants to overwrite the file (if applicable) before returning the filepath. If the path is a directory, it'll just move to that directory instead of trying to save.

Also, got the error working as a popup message.

Still needs a LOT of styling, but looking good!