Fixed issue with multiple listDir tasks.
Now, every time the user changes directory or refreshes, a counter goes up by one. By giving the current count to the listDir task, it can return with the count it was given when run. If that count still matches the current count, it'll update the file list. Otherwise, it does nothing.

In other words, if a directory is taking too long to load, you can go back/up or jump to another directory altogether.